- upgrade commander to 11.1.0 - upgrade @microsoft/api-extractor to 7.39.0 - fixed magda-migrator-registry-aspects - refactor magda-minion-framework - refactor package/acs-cmd - refactor package/arbitraries - refactor package/auth-api-client - refactor package/authentication-plugin-sdk
