+++ magda-minion-framework/src/getWebhookUrl.ts
@@ -1,4 +1,4 @@
-import MinionOptions from "./MinionOptions";
+import MinionOptions from "./MinionOptions.js";
 
 export default function getHookUrl(options: MinionOptions) {
     return `${options.argv.internalUrl}/hook`;
