+++ packages/arbitraries/package.json
@@ -2,22 +2,30 @@
   "name": "@magda/arbitraries",
   "description": "MAGDA jsverify property-based testing arbitraries definitions",
   "version": "2.3.3",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js",
+    "./package.json": "./package.json"
+  },
+  "engines": {
+    "node": ">=16.0.0"
+  },
   "scripts": {
     "prebuild": "rimraf dist tsconfig.tsbuildinfo",
-    "build": "webpack && api-extractor run -l",
+    "build": "yarn build-main && yarn build-types && yarn build-types-bundle",
+    "build-main": "node esbuild.js",
+    "build-types": "tsc --emitDeclarationOnly",
+    "build-types-bundle": "api-extractor run --local",
     "release": "npm publish || echo \"Skip releasing npm package @magda/arbitraties.\""
   },
   "author": "",
   "license": "Apache-2.0",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
   "devDependencies": {
     "@magda/minion-framework": "^2.3.3",
-    "@microsoft/api-extractor": "~7.38.3",
+    "@microsoft/api-extractor": "~7.39.0",
     "ts-loader": "^9.5.1",
     "typescript": "~5.2.2",
-    "webpack": "^5.89.0",
-    "webpack-cli": "^5.1.4"
+    "esbuild": "^0.19.10"
   },
   "dependencies": {
     "jsverify": "^0.8.2"
