+++ packages/authentication-plugin-sdk/api-extractor.json
@@ -60,7 +60,12 @@
      * This would direct API Extractor to embed those types directly in the .d.ts rollup, as if they had been
      * local files for library1.
      */
-    "bundledPackages": ["@magda/typescript-common", "magda-typescript-common"],
+    "bundledPackages": [
+        "@magda/typescript-common",
+        "magda-typescript-common",
+        "@magda/auth-api-client",
+        "urijs"
+    ],
 
     /**
      * Determines how the TypeScript compiler engine will be invoked by API Extractor.
