+++ packages/authentication-plugin-sdk/esbuild.js
@@ -1,0 +1,11 @@
+import * as esbuild from "esbuild";
+
+await esbuild.build({
+    entryPoints: ["./src/index.ts"],
+    bundle: true,
+    platform: "node",
+    target: ["node16"],
+    outdir: "dist",
+    format: "esm",
+    external: ["pg"]
+});
