+++ packages/minion-sdk/package.json
@@ -14,7 +14,7 @@
   "devDependencies": {
     "@magda/minion-framework": "^2.3.3",
     "@magda/typescript-common": "^2.3.3",
-    "@microsoft/api-extractor": "~7.38.3",
+    "@microsoft/api-extractor": "~7.39.0",
     "ts-loader": "^9.5.1",
     "typescript": "~5.2.2",
     "webpack": "^5.89.0",
