+++ scripts/package.json
@@ -52,7 +52,7 @@
     "chalk": "^2.4.1",
     "clear": "^0.1.0",
     "cli-cursor": "^3.1.0",
-    "commander": "^2.17.0",
+    "commander": "^11.1.0",
     "configstore": "^4.0.0",
     "fs-extra": "^2.1.2",
     "fuzzy": "^0.1.3",
